requirements.txt
albumentations==1.1.0
timm==0.4.12 
numpy==1.19.1
opencv_python_headless==4.4.0.44
pytorch_lightning==0.9.0
transformers==3.3.1
scipy==1.5.2
apex==0.9.10dev
scikit_learn==0.23.2
python-dotenv==0.19.2
torch==1.11.0
supervisely==6.23.1
supervisely[apps]==6.23.1
